pyproject.toml
[tool.poetry]
name = "compress_tool"
version = "0.1.0"
description = ""
authors = ["dmitriykodolov <kodolovdv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest-cov = "^3.0.0"
colorama = "^0.4.5"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
